Implement native Off Auto Dynamic in the existing DLSSG dispatcher

+++ OptiScaler/hooks/Streamline_Hooks.h
@@ -149,7 +149,7 @@ class StreamlineHooks
                                        uint32_t& sourceViewport,
                                        uint32_t& mode,
                                        uint32_t& numFramesToGenerate,
-                                       uint32_t& dynamicTargetFrameRate);
+                                       float& dynamicTargetFrameRate);
 
     static void consumeNativeDlssgRequest(uint64_t generation);
